Remove in_transaction? check from failover, as it tries to reconnect
After disconnecting from a database in #hold’s super method,
we are unable to know if the connection has any open transactions.
Attempting to check forces the connection to be re-established,
which when the server is down causes Sequel to go into an infinite
loop.

As a part of this commit, we remove recursion from the #hold
method, replacing it with a loop.